listen feature refactor done
